requirement.txt
streamlit
opencv-python
pillow
ultralytics
